requirements1.txt
pytz==2022.1
requests==2.27.1
selenium==4.1.5
pytest==7.1.2
allure-pytest==2.9.45
allure-python-commons==2.9.45